2.0.0 united game resolution to 800x520 by resize Steam game window (#50)
* change ver 2.0.0

* change method detect screen for Steam on Windows using JNA

* add property requireSpecificClientType to AppMeta

* add texture for profile 800x520

* add warning game window must be clearly visible

* remove WebProfile and SteamProfile

* add issues.bh99bot.com links

* update web UI

* do not generate mini-client upon build project when im.dev exists

* add warning about downloaded wrong zip file when attemp to run bat files on Windows

+++ build.sh
@@ -12,8 +12,10 @@ cp ./target/99bot-*-jar-with-dependencies.jar ./BitHeroes.jar
 touch user-config.properties
 chmod +x ./*.sh
 
-# Generating mini client
-rm -f ./mini-game-on-chrome*.sh
-rm -f mini
-java -jar ./BitHeroes.jar client
-chmod +x ./mini-game-on-chrome*.sh
+if [ ! -f ./im.dev ]; then
+  # Generating mini client
+  rm -f ./mini-game-on-chrome*.sh
+  rm -f mini
+  java -jar ./BitHeroes.jar client
+  chmod +x ./mini-game-on-chrome*.sh
+fi
